Status bar color preservation + Luigi change
Instead of always loading #$3C for a closed inventory, the current color of the status bar is preserved in RAM.

Additionally, a check if playing as Luigi in certain tilesets and color sets which have a red status bar has been added. This will change the bar to green instead.